finishing touches... hopefully
